Conflicts:
	CHANGELOG.md
	src/Pim/Bundle/CatalogBundle/Version.php

+++ src/Pim/Bundle/CatalogBundle/Manager/GroupManager.php
@@ -2,11 +2,13 @@
 
 namespace Pim\Bundle\CatalogBundle\Manager;
 
+use Pim\Bundle\CatalogBundle\Model\GroupInterface;
+use Pim\Bundle\CatalogBundle\Event\GroupEvents;
+use Pim\Component\Resource\Model\RemoverInterface;
+use Pim\Component\Resource\Model\SaverInterface;
+use Symfony\Bridge\Doctrine\RegistryInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\EventDispatcher\GenericEvent;
-use Symfony\Bridge\Doctrine\RegistryInterface;
-use Pim\Bundle\CatalogBundle\Event\GroupEvents;
-use Pim\Bundle\CatalogBundle\Entity\Group;
 
 /**
  * Group manager
@@ -15,7 +17,7 @@ use Pim\Bundle\CatalogBundle\Entity\Group;
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class GroupManager
+class GroupManager implements SaverInterface, RemoverInterface
 {
     /** @var RegistryInterface */
     protected $doctrine;
@@ -64,7 +66,7 @@ class GroupManager
     /**
      * Get available axis
      *
-     * @return \Pim\Bundle\CatalogBundle\Model\AbstractAttribute[]
+     * @return \Pim\Bundle\CatalogBundle\Model\AttributeInterface[]
      */
     public function getAvailableAxis()
     {
@@ -143,28 +145,53 @@ class GroupManager
     }
 
     /**
-     * Removes a group
-     *
-     * @param Group $group
+     * {@inheritdoc}
      */
-    public function remove(Group $group)
+    public function save($group, array $options = [])
     {
+        if (!$group instanceof GroupInterface) {
+            throw new \InvalidArgumentException(
+                sprintf('Expects a "Pim\Bundle\CatalogBundle\Model\GroupInterface", "%s" provided.', get_class($group))
+            );
+        }
+
+        $options = array_merge(['flush' => true], $options);
+        $this->doctrine->getManager()->persist($group);
+        if (true === $options['flush']) {
+            $this->doctrine->getManager()->flush();
+        }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function remove($group, array $options = [])
+    {
+        if (!$group instanceof GroupInterface) {
+            throw new \InvalidArgumentException(
+                sprintf('Expects a "Pim\Bundle\CatalogBundle\Model\GroupInterface", "%s" provided.', get_class($group))
+            );
+        }
+
         $this->eventDispatcher->dispatch(GroupEvents::PRE_REMOVE, new GenericEvent($group));
 
+        $options = array_merge(['flush' => true], $options);
         $em = $this->doctrine->getManager();
         $em->remove($group);
-        $em->flush();
+        if (true === $options['flush']) {
+            $em->flush();
+        }
     }
 
     /**
      * Returns an array containing a limited number of product groups, and the total number of products
      *
-     * @param Group   $group
-     * @param integer $maxResults
+     * @param GroupInterface $group
+     * @param integer        $maxResults
      *
      * @return array
      */
-    public function getProductList(Group $group, $maxResults)
+    public function getProductList(GroupInterface $group, $maxResults)
     {
         $manager = $this->doctrine->getManager();
         $products = $manager
